+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Version 103:
+
+* Boost test matrix fixes
+
+--------------------------------------------------------------------------------
+
 Version 102:
 
 * Section headings in examples